Break up lower/expr.rs into 6 focused modules
expr.rs was 5,404 lines handling all expression lowering in one file.
Split by concern into focused modules that each own a coherent subset:

- iterators.rs (688 lines): iterator chain fusion (.collect, .fold, etc.)
- match_lower.rs (568 lines): match expressions (enum, string, tuple)
- collections.rs (489 lines): Vec/Map construction, JSON encode/decode, clone
- errors.rs (347 lines): try, try-else, map_err lowering
- closures.rs (283 lines): closure synthesis and spawn lowering
- concurrency.rs (265 lines): Shared/Mutex with-block lowering

expr.rs retains the core dispatch (lower_expr match), binary/unary ops,
function calls, method calls, field access, literals, control flow (if),
and block lowering. No single file exceeds 2,644 lines.
