+++ compiler/crates/rask-mir/src/lower/expr.rs
@@ -3,20 +3,19 @@
 //! Expression lowering.
 
 use super::{
-    binop_result_type, is_type_constructor_name, is_variant_name, lower_binop, lower_unaryop,
+    binop_result_type, is_type_constructor_name, lower_binop, lower_unaryop,
     operator_method_to_binop, operator_method_to_unaryop, LoweringError,
     MirLowerer, TypedOperand, HANDLE_NONE_SENTINEL,
 };
 use crate::{
-    operand::MirConst, stmt::ClosureCapture, types::{EnumLayoutId, StructLayoutId}, BlockBuilder,
-    BlockId, FunctionRef, LocalId, MirOperand, MirRValue, MirStmt, MirTerminator, MirType,
+    operand::MirConst, types::{EnumLayoutId, StructLayoutId},
+    BlockId, FunctionRef, MirOperand, MirRValue, MirStmt, MirTerminator, MirType,
 };
 use rask_ast::{
-    expr::{Expr, ExprKind, TryElse, UnaryOp},
+    expr::{Expr, ExprKind, UnaryOp},
     stmt::{Stmt, StmtKind},
     token::{FloatSuffix, IntSuffix},
 };
-use rask_mono::StructLayout;
 
 /// Resolve primitive type associated constants (e.g. i64.MAX, i32.MIN).
 fn primitive_type_constant(type_name: &str, field: &str) -> Option<TypedOperand> {
@@ -2625,627 +2624,6 @@ impl<'a> MirLowerer<'a> {
         Ok((MirOperand::Local(result_local), then_ty))
     }
 
-    /// Match expression lowering (spec L2).
-    ///
-    /// Handles both enum matches (extract tag, switch on discriminant)
-    /// and value matches (switch on scrutinee directly).
-    fn lower_match(
-        &mut self,
-        scrutinee: &Expr,
-        arms: &[rask_ast::expr::MatchArm],
-    ) -> Result<TypedOperand, LoweringError> {
-        use rask_ast::expr::Pattern;
-
-        // Tuple pattern matching: match (a, b) { (p1, p2) => ... }
-        let has_tuple_patterns = arms.iter().any(|a| matches!(&a.pattern, Pattern::Tuple(_)));
-        if has_tuple_patterns {
-            return self.lower_tuple_match(scrutinee, arms);
-        }
-
-        let is_niche = self.is_niche_option_expr(scrutinee);
-        let (scrutinee_op, scrutinee_ty) = self.lower_expr(scrutinee)?;
-
-        // String match: emit chain of string_eq comparisons instead of switch
-        let is_string_match = matches!(scrutinee_ty, MirType::String)
-            || self.ctx.lookup_raw_type(scrutinee.id)
-                .map_or(false, |ty| matches!(ty, rask_types::Type::String));
-        if is_string_match {
-            return self.lower_string_match(scrutinee_op, arms);
-        }
-
-        let is_enum = matches!(scrutinee_ty, MirType::Enum(_));
-
-        // Detect Result/Option via raw type info from type checker
-        let is_result_or_option = if !is_enum {
-            self.ctx.lookup_raw_type(scrutinee.id).map_or(false, |ty| {
-                matches!(ty, rask_types::Type::Result { .. } | rask_types::Type::Option(_))
-            })
-        } else {
-            false
-        };
-
-        // Pattern-based detection: if any arm uses a known variant pattern,
-        // treat the match as tagged dispatch even without type info.
-        let patterns_imply_enum = if !is_enum && !is_result_or_option {
-            arms.iter().any(|arm| match &arm.pattern {
-                Pattern::Constructor { name, .. } => is_variant_name(name),
-                Pattern::Struct { name, .. } => {
-                    self.resolve_pattern_tag(name).is_some()
-                }
-                Pattern::Ident(name) => {
-                    self.resolve_pattern_tag(name).is_some()
-                        || matches!(name.as_str(), "Ok" | "Err" | "Some" | "None")
-                }
-                _ => false,
-            })
-        } else {
-            false
-        };
-        let has_tag = is_enum || is_result_or_option || patterns_imply_enum || is_niche;
-
-        // Extract payload types for Result/Option.
-        // Primary source: type checker's node_types (raw AST types).
-        // Fallback: the already-lowered MIR scrutinee type, which preserves
-        // struct/enum layout IDs even when the type checker entry is missing.
-        let ok_payload_ty = self.extract_payload_type(scrutinee)
-            .or_else(|| match &scrutinee_ty {
-                MirType::Result { ok, .. } => Some(ok.as_ref().clone()),
-                MirType::Option(inner) => Some(inner.as_ref().clone()),
-                _ => None,
-            })
-            .unwrap_or(MirType::I64);
-        let err_payload_ty = self.extract_err_type(scrutinee)
-            .or_else(|| match &scrutinee_ty {
-                MirType::Result { err, .. } => Some(err.as_ref().clone()),
-                _ => None,
-            })
-            .unwrap_or(MirType::I64);
-
-        // Determine the switch value
-        let switch_val = if has_tag {
-            let tag_local = self.emit_option_tag(&scrutinee_op, is_niche);
-            MirOperand::Local(tag_local)
-        } else {
-            scrutinee_op.clone()
-        };
-
-        let merge_block = self.builder.create_block();
-        let arm_blocks: Vec<BlockId> = arms.iter().map(|_| self.builder.create_block()).collect();
-
-        // Build switch cases: map pattern → (value, block)
-        let mut cases: Vec<(u64, BlockId)> = Vec::new();
-        let mut default_block = merge_block;
-
-        for (i, arm) in arms.iter().enumerate() {
-            match &arm.pattern {
-                Pattern::Wildcard => {
-                    default_block = arm_blocks[i];
-                }
-                Pattern::Ident(name) => {
-                    if let Some(tag) = self.resolve_pattern_tag(name) {
-                        cases.push((tag, arm_blocks[i]));
-                    } else if has_tag && is_variant_name(name) {
-                        // Bare variant name (Ok, None, etc.) → use well-known tag
-                        cases.push((self.variant_tag(name) as u64, arm_blocks[i]));
-                    } else {
-                        // Binding variable — acts as default
-                        default_block = arm_blocks[i];
-                    }
-                }
-                Pattern::Constructor { name, .. } => {
-                    if let Some(tag) = self.resolve_pattern_tag(name) {
-                        cases.push((tag, arm_blocks[i]));
-                    } else if has_tag {
-                        cases.push((self.variant_tag(name) as u64, arm_blocks[i]));
-                    } else {
-                        cases.push((i as u64, arm_blocks[i]));
-                    }
-                }
-                Pattern::Literal(lit_expr) => {
-                    if let ExprKind::Int(v, _) = &lit_expr.kind {
-                        cases.push((*v as u64, arm_blocks[i]));
-                    } else if let ExprKind::Bool(b) = &lit_expr.kind {
-                        cases.push((if *b { 1 } else { 0 }, arm_blocks[i]));
-                    } else {
-                        cases.push((i as u64, arm_blocks[i]));
-                    }
-                }
-                Pattern::Struct { name, .. } => {
-                    if let Some(tag) = self.resolve_pattern_tag(name) {
-                        cases.push((tag, arm_blocks[i]));
-                    } else {
-                        cases.push((i as u64, arm_blocks[i]));
-                    }
-                }
-                _ => {
-                    cases.push((i as u64, arm_blocks[i]));
-                }
-            }
-        }
-
-        self.builder.terminate(MirTerminator::Switch {
-            value: switch_val,
-            cases,
-            default: default_block,
-        });
-
-        // Lower each arm
-        let mut result_ty = MirType::Void;
-        let result_local = self.builder.alloc_temp(MirType::I64);
-        for (i, arm) in arms.iter().enumerate() {
-            self.builder.switch_to_block(arm_blocks[i]);
-
-            // Bind pattern variables for enum payloads
-            if has_tag {
-                if let Pattern::Constructor { name, fields } = &arm.pattern {
-                    // Determine field types from enum layout or Result/Option
-                    let variant_fields: Option<Vec<(MirType, u32)>> =
-                        if let MirType::Enum(crate::types::EnumLayoutId(idx)) = &scrutinee_ty {
-                            self.ctx.enum_layouts.get(*idx as usize).and_then(|layout| {
-                                layout.variants.iter().find(|v| v.name == *name).map(|v| {
-                                    v.fields.iter().map(|f| {
-                                        (self.ctx.type_to_mir(&f.ty), f.offset)
-                                    }).collect()
-                                })
-                            })
-                        } else {
-                            None
-                        };
-
-                    for (j, field_pat) in fields.iter().enumerate() {
-                        if let Pattern::Ident(binding) = field_pat {
-                            let field_ty = if let Some(ref vf) = variant_fields {
-                                vf.get(j).map(|(ty, _)| ty.clone()).unwrap_or(MirType::I64)
-                            } else {
-                                // Result/Option: use payload type based on variant
-                                match name.as_str() {
-                                    "Err" => err_payload_ty.clone(),
-                                    _ => ok_payload_ty.clone(),
-                                }
-                            };
-                            let payload_local = self.builder.alloc_local(
-                                binding.clone(), field_ty.clone(),
-                            );
-                            let rvalue = if is_niche {
-                                // Niche: the scrutinee IS the payload
-                                MirRValue::Use(scrutinee_op.clone())
-                            } else {
-                                MirRValue::Field {
-                                    base: scrutinee_op.clone(),
-                                    field_index: j as u32,
-                                    byte_offset: None,
-                                    field_size: None,
-                                }
-                            };
-                            self.builder.push_stmt(MirStmt::Assign {
-                                dst: payload_local,
-                                rvalue,
-                            });
-                            // Set local_type_prefix so method calls on match-bound
-                            // variables get the correct type qualification.
-                            // Derive from MIR type (not node_types which may lack
-                            // monomorphized expression IDs).
-                            let prefix = self.mir_type_name(&field_ty)
-
-                                .or_else(|| {
-                                    // For user-defined enums: look up field type from layout
-                                    if let MirType::Enum(crate::types::EnumLayoutId(idx)) = &scrutinee_ty {
-                                        self.ctx.enum_layouts.get(*idx as usize).and_then(|layout| {
-                                            layout.variants.iter().find(|v| v.name == *name).and_then(|v| {
-                                                v.fields.get(j).and_then(|f| {
-                                                    super::MirContext::type_prefix(&f.ty)
-                                                })
-                                            })
-                                        })
-                                    } else {
-                                        None
-                                    }
-                                })
-                                .or_else(|| {
-                                    // For Result/Option: extract payload MirType from scrutinee
-                                    let payload_mir = match (&scrutinee_ty, name.as_str()) {
-                                        (MirType::Result { err, .. }, "Err") => Some(err.as_ref()),
-                                        (MirType::Result { ok, .. }, _) => Some(ok.as_ref()),
-                                        (MirType::Option(inner), _) => Some(inner.as_ref()),
-                                        _ => None,
-                                    };
-                                    payload_mir.and_then(|t| self.mir_type_name(t))
-                                });
-                            if let Some(p) = prefix {
-                                self.local_type_prefix.insert(binding.clone(), p);
-                            }
-                            self.locals.insert(binding.clone(), (payload_local, field_ty));
-                        }
-                    }
-                } else if let Pattern::Struct { name, fields, .. } = &arm.pattern {
-                    // Named-field destructuring: Shape.Circle { radius }
-                    // Strip enum prefix: "Shape.Circle" → "Circle"
-                    let variant_name = name.rsplit('.').next().unwrap_or(name);
-                    if let MirType::Enum(crate::types::EnumLayoutId(idx)) = &scrutinee_ty {
-                        if let Some(layout) = self.ctx.enum_layouts.get(*idx as usize) {
-                            if let Some(variant) = layout.variants.iter().find(|v| v.name == variant_name) {
-                                for (field_name, field_pat) in fields {
-                                    if let Pattern::Ident(binding) = field_pat {
-                                        if let Some((field_idx, field_layout)) = variant.fields.iter()
-                                            .enumerate()
-                                            .find(|(_, f)| f.name == *field_name)
-                                        {
-                                            let field_ty = self.ctx.type_to_mir(&field_layout.ty);
-                                            let payload_local = self.builder.alloc_local(
-                                                binding.clone(), field_ty.clone(),
-                                            );
-                                            let rvalue = MirRValue::Field {
-                                                base: scrutinee_op.clone(),
-                                                field_index: field_idx as u32,
-                                                byte_offset: None,
-                                                field_size: None,
-                                            };
-                                            self.builder.push_stmt(MirStmt::Assign {
-                                                dst: payload_local,
-                                                rvalue,
-                                            });
-                                            if let Some(p) = self.mir_type_name(&field_ty)
-                                                .or_else(|| super::MirContext::type_prefix(&field_layout.ty))
-                                            {
-                                                self.local_type_prefix.insert(binding.clone(), p);
-                                            }
-                                            self.locals.insert(binding.clone(), (payload_local, field_ty));
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // If the arm has a guard, evaluate it and conditionally skip
-            if let Some(guard_expr) = &arm.guard {
-                let (guard_val, _) = self.lower_expr(guard_expr)?;
-                let guard_fail_block = if i + 1 < arm_blocks.len() {
-                    arm_blocks[i + 1]
-                } else {
-                    default_block
-                };
-                let guard_pass_block = self.builder.create_block();
-                self.builder.terminate(MirTerminator::Branch {
-                    cond: guard_val,
-                    then_block: guard_pass_block,
-                    else_block: guard_fail_block,
-                });
-                self.builder.switch_to_block(guard_pass_block);
-            }
-
-            let (body_val, arm_ty) = self.lower_expr(&arm.body)?;
-            if i == 0 {
-                result_ty = arm_ty;
-            }
-
-            // Only emit Assign+Goto if the arm body didn't already terminate
-            // (e.g., via return, break, or continue).
-            if self.builder.current_block_unterminated() {
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: result_local,
-                    rvalue: MirRValue::Use(body_val),
-                });
-                self.builder.terminate(MirTerminator::Goto {
-                    target: merge_block,
-                });
-            }
-        }
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(result_local), result_ty))
-    }
-
-    /// Lower match on strings: emit a chain of string_eq comparisons.
-    /// `match s { "a" => ..., "b" | "c" => ..., _ => ... }` becomes
-    /// if-else chain: `if string_eq(s, "a") goto arm0 else if ...`
-    fn lower_string_match(
-        &mut self,
-        scrutinee_op: MirOperand,
-        arms: &[rask_ast::expr::MatchArm],
-    ) -> Result<TypedOperand, LoweringError> {
-        use rask_ast::expr::Pattern;
-
-        let merge_block = self.builder.create_block();
-        let arm_blocks: Vec<BlockId> = arms.iter().map(|_| self.builder.create_block()).collect();
-        let result_local = self.builder.alloc_temp(MirType::I64);
-        let mut result_ty = MirType::Void;
-
-        // Find default arm index
-        let default_idx = arms.iter().position(|a| {
-            matches!(&a.pattern, Pattern::Wildcard)
-                || matches!(&a.pattern, Pattern::Ident(n) if !n.starts_with('"'))
-        });
-
-        // Collect string patterns for each arm: Vec<(arm_index, string_literals)>
-        let mut string_arms: Vec<(usize, Vec<String>)> = Vec::new();
-        for (i, arm) in arms.iter().enumerate() {
-            match &arm.pattern {
-                Pattern::Literal(lit) => {
-                    if let ExprKind::String(s) = &lit.kind {
-                        string_arms.push((i, vec![s.clone()]));
-                    }
-                }
-                Pattern::Or(pats) => {
-                    let strs: Vec<String> = pats.iter().filter_map(|p| {
-                        if let Pattern::Literal(lit) = p {
-                            if let ExprKind::String(s) = &lit.kind {
-                                return Some(s.clone());
-                            }
-                        }
-                        None
-                    }).collect();
-                    if !strs.is_empty() {
-                        string_arms.push((i, strs));
-                    }
-                }
-                Pattern::Wildcard | Pattern::Ident(_) => {
-                    // handled as default
-                }
-                _ => {}
-            }
-        }
-
-        let default_block = default_idx.map(|i| arm_blocks[i]).unwrap_or(merge_block);
-
-        // Emit if-else chain
-        for (arm_idx, literals) in &string_arms {
-            for (j, lit) in literals.iter().enumerate() {
-                let eq_result = self.builder.alloc_temp(MirType::Bool);
-                self.builder.push_stmt(MirStmt::Call {
-                    dst: Some(eq_result),
-                    func: FunctionRef::internal("string_eq".to_string()),
-                    args: vec![
-                        scrutinee_op.clone(),
-                        MirOperand::Constant(MirConst::String(lit.clone())),
-                    ],
-                });
-                let next_test = self.builder.create_block();
-                // For Or patterns with multiple literals, on match go to the arm block
-                self.builder.terminate(MirTerminator::Branch {
-                    cond: MirOperand::Local(eq_result),
-                    then_block: arm_blocks[*arm_idx],
-                    else_block: next_test,
-                });
-                self.builder.switch_to_block(next_test);
-                // If last literal and last string arm, fall through to default
-                let _ = j; // used for iteration
-            }
-        }
-        // After all tests, go to default
-        self.builder.terminate(MirTerminator::Goto { target: default_block });
-
-        // Lower each arm body
-        for (i, arm) in arms.iter().enumerate() {
-            self.builder.switch_to_block(arm_blocks[i]);
-
-            // Bind identifier patterns (wildcard variable) to scrutinee
-            if let Pattern::Ident(name) = &arm.pattern {
-                let bind_local = self.builder.alloc_local(name.clone(), MirType::String);
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: bind_local,
-                    rvalue: MirRValue::Use(scrutinee_op.clone()),
-                });
-                self.locals.insert(name.clone(), (bind_local, MirType::String));
-            }
-
-            let (body_val, arm_ty) = self.lower_expr(&arm.body)?;
-            if i == 0 || result_ty == MirType::Void {
-                result_ty = arm_ty;
-            }
-            if self.builder.current_block_unterminated() {
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: result_local,
-                    rvalue: MirRValue::Use(body_val),
-                });
-                self.builder.terminate(MirTerminator::Goto { target: merge_block });
-            }
-        }
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(result_local), result_ty))
-    }
-
-    /// Lower match with tuple patterns: match (a, b) { (p1, p2) => ..., _ => ... }
-    /// Implemented as a chain of conditional branches testing each element.
-    fn lower_tuple_match(
-        &mut self,
-        scrutinee: &Expr,
-        arms: &[rask_ast::expr::MatchArm],
-    ) -> Result<TypedOperand, LoweringError> {
-        use rask_ast::expr::Pattern;
-
-        // Lower tuple elements
-        let tuple_elems: Vec<(MirOperand, MirType)> = if let ExprKind::Tuple(elems) = &scrutinee.kind {
-            let mut result = Vec::new();
-            for elem in elems {
-                result.push(self.lower_expr(elem)?);
-            }
-            result
-        } else {
-            // Single-element fallback
-            vec![self.lower_expr(scrutinee)?]
-        };
-
-        let merge_block = self.builder.create_block();
-        let result_local = self.builder.alloc_temp(MirType::I64);
-        let mut result_ty = MirType::Void;
-
-        // Create test blocks for each arm
-        let arm_test_blocks: Vec<BlockId> = arms.iter().map(|_| self.builder.create_block()).collect();
-        let fallthrough = merge_block;
-
-        // Start with first arm's test block
-        self.builder.terminate(MirTerminator::Goto { target: arm_test_blocks[0] });
-
-        for (i, arm) in arms.iter().enumerate() {
-            self.builder.switch_to_block(arm_test_blocks[i]);
-            let next_arm = if i + 1 < arm_test_blocks.len() {
-                arm_test_blocks[i + 1]
-            } else {
-                fallthrough
-            };
-
-            let sub_patterns = match &arm.pattern {
-                Pattern::Tuple(pats) => pats.clone(),
-                Pattern::Wildcard => {
-                    // Default arm — bind nothing, just fall through to body
-                    vec![]
-                }
-                _ => vec![arm.pattern.clone()],
-            };
-
-            // Check each tuple element against its pattern
-            let body_block = self.builder.create_block();
-            let _current_pass = body_block;
-
-            // Build checks in reverse so we can chain: check0 → check1 → ... → body
-            let mut checks: Vec<(usize, Pattern)> = Vec::new();
-            for (j, pat) in sub_patterns.iter().enumerate() {
-                match pat {
-                    Pattern::Literal(_) => checks.push((j, pat.clone())),
-                    Pattern::Ident(_) | Pattern::Wildcard => {
-                        // Bindings and wildcards always match — handled in body
-                    }
-                    _ => {}
-                }
-            }
-
-            if checks.is_empty() && !matches!(&arm.pattern, Pattern::Wildcard) {
-                // All patterns are bindings — always matches
-                // Bind variables and go to body
-                for (j, pat) in sub_patterns.iter().enumerate() {
-                    if let Pattern::Ident(name) = pat {
-                        if j < tuple_elems.len() {
-                            let (ref elem_op, ref elem_ty) = tuple_elems[j];
-                            let local_id = self.builder.alloc_local(name.clone(), elem_ty.clone());
-                            self.builder.push_stmt(MirStmt::Assign {
-                                dst: local_id,
-                                rvalue: MirRValue::Use(elem_op.clone()),
-                            });
-                            if let Some(prefix) = self.mir_type_name(elem_ty) {
-                                self.local_type_prefix.insert(name.clone(), prefix);
-                            }
-                            self.locals.insert(name.clone(), (local_id, elem_ty.clone()));
-                        }
-                    }
-                }
-            } else if matches!(&arm.pattern, Pattern::Wildcard) {
-                // Wildcard — no checks needed
-            } else {
-                // Emit checks: chain literal comparisons
-                let mut first_check = true;
-                for (j, pat) in &checks {
-                    if let Pattern::Literal(lit_expr) = pat {
-                        let (ref elem_op, _) = tuple_elems[*j];
-                        let (lit_op, _) = self.lower_expr(lit_expr)?;
-
-                        // String comparison
-                        let cmp_local = self.builder.alloc_temp(MirType::I64);
-                        if matches!(&lit_expr.kind, ExprKind::String(_)) {
-                            self.builder.push_stmt(MirStmt::Call {
-                                dst: Some(cmp_local),
-                                func: FunctionRef::internal("string_eq".to_string()),
-                                args: vec![elem_op.clone(), lit_op],
-                            });
-                        } else {
-                            // Integer/bool comparison
-                            self.builder.push_stmt(MirStmt::Assign {
-                                dst: cmp_local,
-                                rvalue: MirRValue::BinaryOp {
-                                    op: crate::BinOp::Eq,
-                                    left: elem_op.clone(),
-                                    right: lit_op,
-                                },
-                            });
-                        }
-
-                        let pass_block = if first_check {
-                            first_check = false;
-                            // More checks or body?
-                            self.builder.create_block()
-                        } else {
-                            self.builder.create_block()
-                        };
-                        self.builder.terminate(MirTerminator::Branch {
-                            cond: MirOperand::Local(cmp_local),
-                            then_block: pass_block,
-                            else_block: next_arm,
-                        });
-                        self.builder.switch_to_block(pass_block);
-                    }
-                }
-
-                // Bind any Ident patterns from the tuple
-                for (j, pat) in sub_patterns.iter().enumerate() {
-                    if let Pattern::Ident(name) = pat {
-                        if j < tuple_elems.len() {
-                            let (ref elem_op, ref elem_ty) = tuple_elems[j];
-                            let local_id = self.builder.alloc_local(name.clone(), elem_ty.clone());
-                            self.builder.push_stmt(MirStmt::Assign {
-                                dst: local_id,
-                                rvalue: MirRValue::Use(elem_op.clone()),
-                            });
-                            if let Some(prefix) = self.mir_type_name(elem_ty) {
-                                self.local_type_prefix.insert(name.clone(), prefix);
-                            }
-                            self.locals.insert(name.clone(), (local_id, elem_ty.clone()));
-                        }
-                    }
-                }
-            }
-
-            // Evaluate guard if present
-            if let Some(guard_expr) = &arm.guard {
-                let (guard_val, _) = self.lower_expr(guard_expr)?;
-                let guard_pass = self.builder.create_block();
-                self.builder.terminate(MirTerminator::Branch {
-                    cond: guard_val,
-                    then_block: guard_pass,
-                    else_block: next_arm,
-                });
-                self.builder.switch_to_block(guard_pass);
-            }
-
-            // Wildcard from non-tuple check still needs goto body
-            if !matches!(&arm.pattern, Pattern::Wildcard) && checks.is_empty() {
-                // Already in body block (bindings-only case)
-            }
-
-            // Lower body
-            let (body_val, arm_ty) = self.lower_expr(&arm.body)?;
-            if i == 0 { result_ty = arm_ty; }
-
-            if self.builder.current_block_unterminated() {
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: result_local,
-                    rvalue: MirRValue::Use(body_val),
-                });
-                self.builder.terminate(MirTerminator::Goto { target: merge_block });
-            }
-        }
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(result_local), result_ty))
-    }
-
-    /// Resolve enum variant name to its tag value from the layout.
-    /// Handles "Color.Red" → 0, "Color.Green" → 1, etc.
-    fn resolve_pattern_tag(&self, name: &str) -> Option<u64> {
-        // Pattern name could be "Color.Red" (qualified) or just "Red" (variant only)
-        let parts: Vec<&str> = name.splitn(2, '.').collect();
-        let (enum_name, variant_name) = if parts.len() == 2 {
-            (parts[0], parts[1])
-        } else {
-            return None;
-        };
-        let (_, layout) = self.ctx.find_enum(enum_name)?;
-        let variant = layout.variants.iter().find(|v| v.name == variant_name)?;
-        Some(variant.tag)
-    }
-
     /// Block expression: lower each statement, last expression is the value.
     pub(super) fn lower_block(&mut self, stmts: &[Stmt]) -> Result<TypedOperand, LoweringError> {
         let mut last_val = MirOperand::Constant(MirConst::Int(0));
@@ -3263,2142 +2641,4 @@ impl<'a> MirLowerer<'a> {
         }
         Ok((last_val, last_ty))
     }
-
-    /// Closure lowering: synthesize a separate MIR function for the body,
-    /// build the environment, and emit ClosureCreate in the enclosing function.
-    fn lower_closure(
-        &mut self,
-        params: &[rask_ast::expr::ClosureParam],
-        ret_ty: Option<&str>,
-        body: &Expr,
-    ) -> Result<TypedOperand, LoweringError> {
-        // 1. Collect free variables (captures from enclosing scope)
-        let free_vars = self.collect_free_vars(body, params);
-
-        // 2. Generate unique name for the closure function
-        let closure_name = format!("{}__closure_{}", self.parent_name, self.closure_counter);
-        self.closure_counter += 1;
-
-        // 3. Build the closure environment layout
-        let mut captures = Vec::new();
-        let mut env_offset = 0u32;
-        for (_name, local_id, ty) in &free_vars {
-            let size = ty.size();
-            // 8-byte alignment
-            let aligned_offset = (env_offset + 7) & !7;
-            captures.push(ClosureCapture {
-                local_id: *local_id,
-                offset: aligned_offset,
-                size,
-            });
-            env_offset = aligned_offset + size;
-        }
-
-        // 4. Synthesize a MIR function for the closure body.
-        //    Signature: fn closure_name(env_ptr: ptr, params...) -> ret
-        let closure_ret = ret_ty
-            .map(|s| self.ctx.resolve_type_str(s))
-            .unwrap_or(MirType::I64);
-        let mut closure_builder = BlockBuilder::new(closure_name.clone(), closure_ret.clone());
-
-        // env_ptr is the implicit first parameter
-        let env_param_id = closure_builder.add_param("__env".to_string(), MirType::Ptr);
-
-        // Add explicit closure parameters
-        let mut closure_locals = std::collections::HashMap::new();
-        for param in params {
-            let param_ty = param.ty.as_deref()
-                .map(|s| self.ctx.resolve_type_str(s))
-                .unwrap_or(MirType::I64);
-            let param_id = closure_builder.add_param(param.name.clone(), param_ty.clone());
-            closure_locals.insert(param.name.clone(), (param_id, param_ty.clone()));
-            // Set type prefix for closure params so method calls inside qualify correctly.
-            // local_type_prefix isn't swapped during closure lowering, so this is visible in body.
-            if let Some(prefix) = self.mir_type_name(&param_ty) {
-                self.local_type_prefix.insert(param.name.clone(), prefix);
-            } else if let Some(ty_str) = param.ty.as_deref() {
-                if let Some(prefix) = super::type_prefix_from_str(ty_str) {
-                    self.local_type_prefix.insert(param.name.clone(), prefix);
-                }
-            }
-        }
-
-        // Emit LoadCapture for each free variable
-        for (i, (name, _outer_id, ty)) in free_vars.iter().enumerate() {
-            let cap = &captures[i];
-            let local_id = closure_builder.alloc_local(name.clone(), ty.clone());
-            closure_builder.push_stmt(MirStmt::LoadCapture {
-                dst: local_id,
-                env_ptr: env_param_id,
-                offset: cap.offset,
-            });
-            closure_locals.insert(name.clone(), (local_id, ty.clone()));
-        }
-
-        // Lower the closure body using a temporary lowerer
-        {
-            let saved_builder = std::mem::replace(&mut self.builder, closure_builder);
-            let saved_locals = std::mem::replace(&mut self.locals, closure_locals);
-            let saved_loop_stack = std::mem::take(&mut self.loop_stack);
-
-            let body_result = self.lower_expr(body);
-
-            // Restore parent state
-            closure_builder = std::mem::replace(&mut self.builder, saved_builder);
-            self.locals = saved_locals;
-            self.loop_stack = saved_loop_stack;
-
-            let (body_val, _body_ty) = body_result?;
-
-            // Add implicit return of body value if block is unterminated
-            if closure_builder.current_block_unterminated() {
-                closure_builder.terminate(MirTerminator::Return {
-                    value: Some(body_val),
-                });
-            }
-        }
-
-        let closure_fn = closure_builder.finish();
-
-        // Register the closure function signature for return type lookup
-        self.func_sigs.insert(closure_name.clone(), super::FuncSig {
-            ret_ty: closure_ret,
-        });
-
-        self.synthesized_functions.push(closure_fn);
-
-        // 5. In the parent function, emit ClosureCreate (heap by default;
-        //    the closure optimization pass downgrades to stack when safe).
-        let result_local = self.builder.alloc_temp(MirType::Ptr);
-        self.builder.push_stmt(MirStmt::ClosureCreate {
-            dst: result_local,
-            func_name: closure_name,
-            captures,
-            heap: true,
-        });
-
-        Ok((MirOperand::Local(result_local), MirType::Ptr))
-    }
-
-    /// Extract the inner struct size from a generic type name like "Channel<LogEntry>".
-    /// Returns 8 (scalar size) if the inner type is not a known struct.
-    fn generic_inner_struct_size(&self, generic_name: &str) -> i64 {
-        let inner = generic_name.split('<').nth(1)
-            .and_then(|s| s.strip_suffix('>'));
-        if let Some(type_name) = inner {
-            if let Some((_, layout)) = self.ctx.find_struct(type_name) {
-                return layout.size as i64;
-            }
-        }
-        8 // scalar default
-    }
-
-
-    /// Extract the inner type name from a Shared variable expression.
-    /// Tries type checker info first, falls back to local_type_prefix context.
-    fn resolve_shared_inner_type_name(&self, object: &Expr) -> Option<String> {
-        // Try type checker: Shared<Database> → "Database"
-        if let Some(raw_ty) = self.ctx.lookup_raw_type(object.id) {
-            if let rask_types::Type::UnresolvedGeneric { args, .. } = raw_ty {
-                if let Some(rask_types::GenericArg::Type(inner)) = args.first() {
-                    if let rask_types::Type::UnresolvedNamed(name) = inner.as_ref() {
-                        return Some(name.clone());
-                    }
-                    // Also handle Named types
-                    if let Some(prefix) = super::MirContext::type_prefix(inner) {
-                        return Some(prefix);
-                    }
-                }
-            }
-        }
-        // Fallback: parse inner type from stored full type annotation (e.g. "Shared<Database>" → "Database")
-        if let ExprKind::Ident(var_name) = &object.kind {
-            if let Some(full_type) = self.local_full_type.get(var_name) {
-                let inner = full_type.split('<').nth(1)
-                    .and_then(|s| s.strip_suffix('>'));
-                if let Some(name) = inner {
-                    return Some(name.to_string());
-                }
-            }
-        }
-        None
-    }
-
-    /// Lower `with shared.read() as d { body }` / `with shared.write() as d { body }`.
-    ///
-    /// Synthesizes a closure from the body with `d` as a parameter, then calls
-    /// Shared_read(handle, closure) or Shared_write(handle, closure).
-    /// The C runtime acquires the lock, passes data to the closure, and releases.
-    fn lower_shared_with_block(
-        &mut self,
-        object: &Expr,
-        method: &str,
-        binding_name: &str,
-        body: &[rask_ast::stmt::Stmt],
-    ) -> Result<TypedOperand, LoweringError> {
-        // 1. Evaluate the Shared object to get the handle
-        let (shared_op, _) = self.lower_expr(object)?;
-
-        // 2. Collect free variables from the body, excluding the binding name
-        let mut free_vars = Vec::new();
-        let mut seen = std::collections::HashSet::new();
-        let mut bound = std::collections::HashSet::new();
-        bound.insert(binding_name.to_string());
-        self.walk_free_vars_block(body, &bound, &mut seen, &mut free_vars);
-
-        // 3. Build closure environment layout
-        let closure_name = format!("{}__with_{}", self.parent_name, self.closure_counter);
-        self.closure_counter += 1;
-
-        let mut captures = Vec::new();
-        let mut env_offset = 0u32;
-        for (_name, local_id, ty) in &free_vars {
-            let size = ty.size();
-            let aligned_offset = (env_offset + 7) & !7;
-            captures.push(ClosureCapture {
-                local_id: *local_id,
-                offset: aligned_offset,
-                size,
-            });
-            env_offset = aligned_offset + size;
-        }
-
-        // 4. Synthesize closure: fn(env_ptr: ptr, data: i64) -> i64
-        let mut closure_builder = BlockBuilder::new(closure_name.clone(), MirType::I64);
-
-        // env_ptr is the implicit first parameter
-        let env_param_id = closure_builder.add_param("__env".to_string(), MirType::Ptr);
-
-        // Resolve the Shared's inner type: struct layout for field offsets, prefix for method dispatch.
-        let mut data_param_ty = MirType::I64;
-        let inner_type_name = self.resolve_shared_inner_type_name(object);
-        if let Some(ref type_name) = inner_type_name {
-            if let Some((layout_idx, _)) = self.ctx.find_struct(type_name) {
-                data_param_ty = MirType::Struct(StructLayoutId(layout_idx));
-            }
-            self.local_type_prefix.insert(binding_name.to_string(), type_name.clone());
-        }
-
-        // data parameter — the Shared's inner data, passed by the C runtime
-        let data_param_id = closure_builder.add_param(binding_name.to_string(), data_param_ty.clone());
-
-        let mut closure_locals = std::collections::HashMap::new();
-        closure_locals.insert(binding_name.to_string(), (data_param_id, data_param_ty));
-
-        // Emit LoadCapture for each free variable
-        for (i, (name, _outer_id, ty)) in free_vars.iter().enumerate() {
-            let cap = &captures[i];
-            let local_id = closure_builder.alloc_local(name.clone(), ty.clone());
-            closure_builder.push_stmt(MirStmt::LoadCapture {
-                dst: local_id,
-                env_ptr: env_param_id,
-                offset: cap.offset,
-            });
-            closure_locals.insert(name.clone(), (local_id, ty.clone()));
-        }
-
-        // 5. Lower the body
-        let body_result;
-        {
-            let saved_builder = std::mem::replace(&mut self.builder, closure_builder);
-            let saved_locals = std::mem::replace(&mut self.locals, closure_locals);
-            let saved_loop_stack = std::mem::take(&mut self.loop_stack);
-
-            body_result = self.lower_block(body);
-
-            closure_builder = std::mem::replace(&mut self.builder, saved_builder);
-            self.locals = saved_locals;
-            self.loop_stack = saved_loop_stack;
-        }
-
-        let (body_val, _) = body_result?;
-
-        if closure_builder.current_block_unterminated() {
-            closure_builder.terminate(MirTerminator::Return {
-                value: Some(body_val),
-            });
-        }
-
-        let closure_fn = closure_builder.finish();
-        self.func_sigs.insert(closure_name.clone(), super::FuncSig {
-            ret_ty: MirType::I64,
-        });
-        self.synthesized_functions.push(closure_fn);
-
-        // 6. In the parent: create closure, call Shared_read/Shared_write
-        let closure_local = self.builder.alloc_temp(MirType::Ptr);
-        self.builder.push_stmt(MirStmt::ClosureCreate {
-            dst: closure_local,
-            func_name: closure_name,
-            captures,
-            heap: false,
-        });
-
-        let func_name = if method == "read" {
-            "Shared_read".to_string()
-        } else {
-            "Shared_write".to_string()
-        };
-
-        let result_local = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(result_local),
-            func: FunctionRef::internal(func_name),
-            args: vec![shared_op, MirOperand::Local(closure_local)],
-        });
-
-        Ok((MirOperand::Local(result_local), MirType::I64))
-    }
-
-    /// Lower `with mutex as v { body }`.
-    ///
-    /// Synthesizes a closure from the body and calls Mutex_lock(handle, closure).
-    /// The C runtime acquires the lock, passes data to the closure, and releases.
-    fn lower_mutex_with_block(
-        &mut self,
-        object: &Expr,
-        binding_name: &str,
-        body: &[rask_ast::stmt::Stmt],
-    ) -> Result<TypedOperand, LoweringError> {
-        let (mutex_op, _) = self.lower_expr(object)?;
-
-        let mut free_vars = Vec::new();
-        let mut seen = std::collections::HashSet::new();
-        let mut bound = std::collections::HashSet::new();
-        bound.insert(binding_name.to_string());
-        self.walk_free_vars_block(body, &bound, &mut seen, &mut free_vars);
-
-        let closure_name = format!("{}__with_{}", self.parent_name, self.closure_counter);
-        self.closure_counter += 1;
-
-        let mut captures = Vec::new();
-        let mut env_offset = 0u32;
-        for (_name, local_id, ty) in &free_vars {
-            let size = ty.size();
-            let aligned_offset = (env_offset + 7) & !7;
-            captures.push(ClosureCapture {
-                local_id: *local_id,
-                offset: aligned_offset,
-                size,
-            });
-            env_offset = aligned_offset + size;
-        }
-
-        let mut closure_builder = BlockBuilder::new(closure_name.clone(), MirType::I64);
-        let env_param_id = closure_builder.add_param("__env".to_string(), MirType::Ptr);
-
-        // Resolve inner type — reuse Shared's resolver since Mutex<T> has the same shape
-        let mut data_param_ty = MirType::I64;
-        let inner_type_name = self.resolve_shared_inner_type_name(object);
-        if let Some(ref type_name) = inner_type_name {
-            if let Some((layout_idx, _)) = self.ctx.find_struct(type_name) {
-                data_param_ty = MirType::Struct(StructLayoutId(layout_idx));
-            }
-            self.local_type_prefix.insert(binding_name.to_string(), type_name.clone());
-        }
-
-        let data_param_id = closure_builder.add_param(binding_name.to_string(), data_param_ty.clone());
-
-        let mut closure_locals = std::collections::HashMap::new();
-        closure_locals.insert(binding_name.to_string(), (data_param_id, data_param_ty));
-
-        for (i, (name, _outer_id, ty)) in free_vars.iter().enumerate() {
-            let cap = &captures[i];
-            let local_id = closure_builder.alloc_local(name.clone(), ty.clone());
-            closure_builder.push_stmt(MirStmt::LoadCapture {
-                dst: local_id,
-                env_ptr: env_param_id,
-                offset: cap.offset,
-            });
-            closure_locals.insert(name.clone(), (local_id, ty.clone()));
-        }
-
-        let body_result;
-        {
-            let saved_builder = std::mem::replace(&mut self.builder, closure_builder);
-            let saved_locals = std::mem::replace(&mut self.locals, closure_locals);
-            let saved_loop_stack = std::mem::take(&mut self.loop_stack);
-
-            body_result = self.lower_block(body);
-
-            closure_builder = std::mem::replace(&mut self.builder, saved_builder);
-            self.locals = saved_locals;
-            self.loop_stack = saved_loop_stack;
-        }
-
-        let (body_val, _) = body_result?;
-
-        if closure_builder.current_block_unterminated() {
-            closure_builder.terminate(MirTerminator::Return {
-                value: Some(body_val),
-            });
-        }
-
-        let closure_fn = closure_builder.finish();
-        self.func_sigs.insert(closure_name.clone(), super::FuncSig {
-            ret_ty: MirType::I64,
-        });
-        self.synthesized_functions.push(closure_fn);
-
-        let closure_local = self.builder.alloc_temp(MirType::Ptr);
-        self.builder.push_stmt(MirStmt::ClosureCreate {
-            dst: closure_local,
-            func_name: closure_name,
-            captures,
-            heap: false,
-        });
-
-        let result_local = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(result_local),
-            func: FunctionRef::internal("Mutex_lock".to_string()),
-            args: vec![mutex_op, MirOperand::Local(closure_local)],
-        });
-
-        Ok((MirOperand::Local(result_local), MirType::I64))
-    }
-
-    /// Spawn lowering: synthesize a closure function from the body block,
-    /// emit ClosureCreate + Call to rask_closure_spawn.
-    ///
-    /// The closure function has signature `fn(env_ptr: ptr)` — no params, void return.
-    /// rask_closure_spawn extracts func_ptr from the closure, spawns an OS thread,
-    /// and frees the closure allocation when the task completes.
-    fn lower_spawn(
-        &mut self,
-        body: &[Stmt],
-    ) -> Result<TypedOperand, LoweringError> {
-        // 1. Collect free variables from the spawn body block
-        let free_vars = self.collect_free_vars_block(body);
-
-        // 2. Generate unique name for the spawn function
-        let spawn_name = format!("{}__spawn_{}", self.parent_name, self.closure_counter);
-        self.closure_counter += 1;
-
-        // 3. Build the closure environment layout
-        let mut captures = Vec::new();
-        let mut env_offset = 0u32;
-        for (_name, local_id, ty) in &free_vars {
-            let size = ty.size();
-            let aligned_offset = (env_offset + 7) & !7;
-            captures.push(ClosureCapture {
-                local_id: *local_id,
-                offset: aligned_offset,
-                size,
-            });
-            env_offset = aligned_offset + size;
-        }
-
-        // 4. Synthesize a MIR function for the spawn body.
-        //    Signature: fn spawn_name(env_ptr: ptr) -> void
-        let mut spawn_builder = BlockBuilder::new(spawn_name.clone(), MirType::Void);
-
-        // env_ptr is the implicit first parameter (standard closure convention)
-        let env_param_id = spawn_builder.add_param("__env".to_string(), MirType::Ptr);
-
-        // Emit LoadCapture for each free variable
-        let mut spawn_locals = std::collections::HashMap::new();
-        for (i, (name, _outer_id, ty)) in free_vars.iter().enumerate() {
-            let cap = &captures[i];
-            let local_id = spawn_builder.alloc_local(name.clone(), ty.clone());
-            spawn_builder.push_stmt(MirStmt::LoadCapture {
-                dst: local_id,
-                env_ptr: env_param_id,
-                offset: cap.offset,
-            });
-            spawn_locals.insert(name.clone(), (local_id, ty.clone()));
-        }
-
-        // Lower the body statements using a temporary lowerer
-        {
-            let saved_builder = std::mem::replace(&mut self.builder, spawn_builder);
-            let saved_locals = std::mem::replace(&mut self.locals, spawn_locals);
-            let saved_loop_stack = std::mem::take(&mut self.loop_stack);
-
-            let mut body_result = Ok(());
-            for stmt in body {
-                if let Err(e) = self.lower_stmt(stmt) {
-                    body_result = Err(e);
-                    break;
-                }
-            }
-
-            // Restore parent state
-            spawn_builder = std::mem::replace(&mut self.builder, saved_builder);
-            self.locals = saved_locals;
-            self.loop_stack = saved_loop_stack;
-
-            body_result?;
-
-            // Add implicit void return if unterminated
-            if spawn_builder.current_block_unterminated() {
-                spawn_builder.terminate(MirTerminator::Return { value: None });
-            }
-        }
-
-        let spawn_fn = spawn_builder.finish();
-
-        // Try the state machine transform for yield-point-containing spawns
-        if let Some(sm_result) = crate::transform::state_machine::transform(&spawn_fn) {
-            // Register the poll function
-            let poll_name = sm_result.poll_fn.name.clone();
-            self.func_sigs.insert(poll_name.clone(), super::FuncSig {
-                ret_ty: MirType::I32,
-            });
-            self.synthesized_functions.push(sm_result.poll_fn);
-
-            // Allocate state struct, init tag = 0, store captured vars
-            let state_ptr = self.builder.alloc_temp(MirType::Ptr);
-            let state_size_val = sm_result.state_size as i64;
-            self.builder.push_stmt(MirStmt::Call {
-                dst: Some(state_ptr),
-                func: FunctionRef::internal("rask_alloc".to_string()),
-                args: vec![MirOperand::Constant(crate::operand::MirConst::Int(state_size_val))],
-            });
-
-            // Store state_tag = 0
-            self.builder.push_stmt(MirStmt::Store {
-                addr: state_ptr,
-                offset: 0,
-                value: MirOperand::Constant(crate::operand::MirConst::Int(0)),
-                store_size: None,
-            });
-
-            // Store captured variables into the state struct.
-            // capture_stores maps (env_offset → state_offset). Match each
-            // entry to the ClosureCapture with the same env_offset to find
-            // the parent's local_id.
-            for &(env_offset, state_offset) in &sm_result.capture_stores {
-                if let Some(cap) = captures.iter().find(|c| c.offset == env_offset) {
-                    self.builder.push_stmt(MirStmt::Store {
-                        addr: state_ptr,
-                        offset: state_offset,
-                        value: MirOperand::Local(cap.local_id),
-                        store_size: None,
-                    });
-                }
-            }
-
-            // Emit: rask_green_spawn(poll_fn, state_ptr, state_size)
-            let poll_fn_ptr = self.builder.alloc_temp(MirType::Ptr);
-            self.builder.push_stmt(MirStmt::Assign {
-                dst: poll_fn_ptr,
-                rvalue: MirRValue::Use(MirOperand::Constant(
-                    crate::operand::MirConst::String(poll_name),
-                )),
-            });
-
-            let handle_local = self.builder.alloc_temp(MirType::Ptr);
-            self.builder.push_stmt(MirStmt::Call {
-                dst: Some(handle_local),
-                func: FunctionRef::internal("rask_green_spawn".to_string()),
-                args: vec![
-                    MirOperand::Local(poll_fn_ptr),
-                    MirOperand::Local(state_ptr),
-                    MirOperand::Constant(crate::operand::MirConst::Int(state_size_val)),
-                ],
-            });
-
-            Ok((MirOperand::Local(handle_local), MirType::Ptr))
-        } else {
-            // No yield points — use the closure bridge
-            // Register the spawn function signature
-            self.func_sigs.insert(spawn_name.clone(), super::FuncSig {
-                ret_ty: MirType::Void,
-            });
-            self.synthesized_functions.push(spawn_fn);
-
-            // Emit ClosureCreate + rask_green_closure_spawn call
-            let closure_local = self.builder.alloc_temp(MirType::Ptr);
-            self.builder.push_stmt(MirStmt::ClosureCreate {
-                dst: closure_local,
-                func_name: spawn_name,
-                captures,
-                heap: true,
-            });
-
-            let handle_local = self.builder.alloc_temp(MirType::Ptr);
-            self.builder.push_stmt(MirStmt::Call {
-                dst: Some(handle_local),
-                func: FunctionRef::internal("spawn".to_string()),
-                args: vec![MirOperand::Local(closure_local)],
-            });
-
-            Ok((MirOperand::Local(handle_local), MirType::Ptr))
-        }
-    }
-
-    /// Collect free variables from a block of statements (no params to bind).
-    fn collect_free_vars_block(
-        &self,
-        body: &[Stmt],
-    ) -> Vec<(String, LocalId, MirType)> {
-        let mut free = Vec::new();
-        let mut seen = std::collections::HashSet::new();
-        let bound = std::collections::HashSet::new();
-        self.walk_free_vars_block(body, &bound, &mut seen, &mut free);
-        free
-    }
-
-    /// Try expression lowering (spec L3).
-    fn lower_try(&mut self, inner: &Expr) -> Result<TypedOperand, LoweringError> {
-        let (result, result_ty) = self.lower_expr(inner)?;
-
-        let tag_local = self.builder.alloc_temp(MirType::U8);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: tag_local,
-            rvalue: MirRValue::EnumTag {
-                value: result.clone(),
-            },
-        });
-
-        let ok_block = self.builder.create_block();
-        let err_block = self.builder.create_block();
-        let merge_block = self.builder.create_block();
-
-        self.builder.terminate(MirTerminator::Branch {
-            cond: MirOperand::Local(tag_local),
-            then_block: err_block,
-            else_block: ok_block,
-        });
-
-        // Err path — extract the error payload and return it
-        self.builder.switch_to_block(err_block);
-        let err_ty = self.extract_err_type(inner)
-            .or_else(|| match &result_ty {
-                MirType::Result { err, .. } => Some(err.as_ref().clone()),
-                _ => None,
-            })
-            .unwrap_or(MirType::I64);
-        let err_val = self.builder.alloc_temp(err_ty);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: err_val,
-            rvalue: MirRValue::Field {
-                base: result.clone(),
-                field_index: 0,
-                byte_offset: None,
-                field_size: None,
-            },
-        });
-        self.builder.terminate(MirTerminator::Return {
-            value: Some(MirOperand::Local(err_val)),
-        });
-
-        // Ok path
-        self.builder.switch_to_block(ok_block);
-        // Infer Ok payload type: type checker → MIR type → walk AST for base call
-        let ok_ty = self.extract_payload_type(inner)
-            .or_else(|| match &result_ty {
-                MirType::Result { ok, .. } => Some(ok.as_ref().clone()),
-                _ => None,
-            })
-            .or_else(|| {
-                // Walk through method chains (e.g. expr.map_err(...)) to find
-                // the base function call and extract Ok type from its return sig.
-                let mut expr = inner;
-                loop {
-                    match &expr.kind {
-                        ExprKind::MethodCall { object, method, .. } => {
-                            // Module.function() pattern (e.g. fs.read_file)
-                            if let ExprKind::Ident(mod_name) = &object.kind {
-                                if super::is_type_constructor_name(mod_name) {
-                                    let func_name = format!("{}_{}", mod_name, method);
-                                    let ret = self.func_sigs.get(&func_name)
-                                        .map(|s| s.ret_ty.clone())
-                                        .unwrap_or_else(|| super::stdlib_return_mir_type(&func_name));
-                                    return match ret {
-                                        MirType::Result { ok, .. } => Some(*ok),
-                                        MirType::Option(inner) => Some(*inner),
-                                        _ => None,
-                                    };
-                                }
-                            }
-                            expr = object;
-                        }
-                        ExprKind::Call { func, .. } => {
-                            let name = match &func.kind {
-                                ExprKind::Ident(n) => n.clone(),
-                                ExprKind::Field { object: o, field: f } => {
-                                    if let ExprKind::Ident(mod_name) = &o.kind {
-                                        format!("{}_{}", mod_name, f)
-                                    } else { break; }
-                                }
-                                _ => break,
-                            };
-                            let ret = self.func_sigs.get(&name)
-                                .map(|s| s.ret_ty.clone())
-                                .unwrap_or_else(|| super::stdlib_return_mir_type(&name));
-                            return match ret {
-                                MirType::Result { ok, .. } => Some(*ok),
-                                MirType::Option(inner) => Some(*inner),
-                                _ => None,
-                            };
-                        }
-                        _ => break,
-                    }
-                }
-                None
-            })
-            .unwrap_or(MirType::I64);
-        let ok_val = self.builder.alloc_temp(ok_ty.clone());
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: ok_val,
-            rvalue: MirRValue::Field {
-                base: result,
-                field_index: 0,
-                byte_offset: None,
-                field_size: None,
-            },
-        });
-        self.builder.terminate(MirTerminator::Goto {
-            target: merge_block,
-        });
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(ok_val), ok_ty))
-    }
-
-    /// Try-else expression: `try expr else |e| { transform(e) }`
-    ///
-    /// Like lower_try but the error path evaluates the else body (with the
-    /// error bound to the parameter) and returns the transformed value.
-    fn lower_try_else(&mut self, inner: &Expr, try_else: &TryElse) -> Result<TypedOperand, LoweringError> {
-        let (result, result_ty) = self.lower_expr(inner)?;
-
-        let tag_local = self.builder.alloc_temp(MirType::U8);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: tag_local,
-            rvalue: MirRValue::EnumTag {
-                value: result.clone(),
-            },
-        });
-
-        let ok_block = self.builder.create_block();
-        let err_block = self.builder.create_block();
-        let merge_block = self.builder.create_block();
-
-        self.builder.terminate(MirTerminator::Branch {
-            cond: MirOperand::Local(tag_local),
-            then_block: err_block,
-            else_block: ok_block,
-        });
-
-        // Err path — bind error to param, evaluate else body, return transformed error
-        self.builder.switch_to_block(err_block);
-        let err_ty = self.extract_err_type(inner)
-            .or_else(|| match &result_ty {
-                MirType::Result { err, .. } => Some(err.as_ref().clone()),
-                _ => None,
-            })
-            .unwrap_or(MirType::I64);
-        let err_val = self.builder.alloc_temp(err_ty.clone());
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: err_val,
-            rvalue: MirRValue::Field {
-                base: result.clone(),
-                field_index: 0,
-                byte_offset: None,
-                field_size: None,
-            },
-        });
-
-        // Bind the error to the else clause's parameter name
-        let err_binding = &try_else.error_binding;
-        self.locals.insert(err_binding.clone(), (err_val, err_ty));
-
-        // Lower the else body — produces the transformed error value
-        let (transformed_op, _transformed_ty) = self.lower_expr(&try_else.body)?;
-
-        self.builder.terminate(MirTerminator::Return {
-            value: Some(transformed_op),
-        });
-
-        // Ok path — extract payload
-        self.builder.switch_to_block(ok_block);
-        let ok_ty = self.extract_payload_type(inner)
-            .or_else(|| match &result_ty {
-                MirType::Result { ok, .. } => Some(ok.as_ref().clone()),
-                _ => None,
-            })
-            .unwrap_or(MirType::I64);
-        let ok_val = self.builder.alloc_temp(ok_ty.clone());
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: ok_val,
-            rvalue: MirRValue::Field {
-                base: result,
-                field_index: 0,
-                byte_offset: None,
-                field_size: None,
-            },
-        });
-        self.builder.terminate(MirTerminator::Goto {
-            target: merge_block,
-        });
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(ok_val), ok_ty))
-    }
-
-    /// Inline expansion of `result.map_err(|e| transform(e))`.
-    ///
-    /// Checks the result tag: Ok (tag=0) passes through unchanged,
-    /// Err (tag=1) extracts the payload, calls the closure, and rewraps.
-    fn lower_map_err(
-        &mut self,
-        result_op: MirOperand,
-        result_ty: &MirType,
-        closure_expr: &Expr,
-    ) -> Result<TypedOperand, LoweringError> {
-        // Lower the closure to get a callable local
-        let (closure_op, _) = self.lower_expr(closure_expr)?;
-        let closure_local = self.builder.alloc_temp(MirType::Ptr);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: closure_local,
-            rvalue: MirRValue::Use(closure_op),
-        });
-
-        // Extract tag
-        let tag_local = self.builder.alloc_temp(MirType::U8);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: tag_local,
-            rvalue: MirRValue::EnumTag { value: result_op.clone() },
-        });
-
-        let ok_block = self.builder.create_block();
-        let err_block = self.builder.create_block();
-        let merge_block = self.builder.create_block();
-
-        // Preserve Result type so codegen allocates a stack slot and uses
-        // correct field offsets (tag at 0, payload at 8).
-        let out_ty = result_ty.clone();
-        let out = self.builder.alloc_temp(out_ty.clone());
-
-        // tag=0 → Ok, tag=1 → Err
-        self.builder.terminate(MirTerminator::Branch {
-            cond: MirOperand::Local(tag_local),
-            then_block: err_block,
-            else_block: ok_block,
-        });
-
-        // Ok path: pass through unchanged
-        self.builder.switch_to_block(ok_block);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: out,
-            rvalue: MirRValue::Use(result_op.clone()),
-        });
-        self.builder.terminate(MirTerminator::Goto { target: merge_block });
-
-        // Err path: extract payload, call closure, wrap as Err
-        self.builder.switch_to_block(err_block);
-        let err_payload = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: err_payload,
-            rvalue: MirRValue::Field { base: result_op, field_index: 0, byte_offset: None, field_size: None },
-        });
-        let new_err = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::ClosureCall {
-            dst: Some(new_err),
-            closure: closure_local,
-            args: vec![MirOperand::Local(err_payload)],
-        });
-        // Construct Err(new_err): tag=1, payload=new_err
-        self.builder.push_stmt(MirStmt::Store {
-            addr: out,
-            offset: 0,
-            value: MirOperand::Constant(MirConst::Int(1)),
-            store_size: None,
-        });
-        self.builder.push_stmt(MirStmt::Store {
-            addr: out,
-            offset: 8,
-            value: MirOperand::Local(new_err),
-            store_size: None,
-        });
-        self.builder.terminate(MirTerminator::Goto { target: merge_block });
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(out), out_ty))
-    }
-
-    /// Inline expansion of `result.map_err(VariantConstructor)`.
-    ///
-    /// Same logic as the closure version: Ok passes through, Err extracts
-    /// the payload and wraps it with the variant constructor.
-    fn lower_map_err_constructor(
-        &mut self,
-        result_op: MirOperand,
-        result_ty: &MirType,
-        constructor_name: &str,
-    ) -> Result<TypedOperand, LoweringError> {
-        // Extract tag
-        let tag_local = self.builder.alloc_temp(MirType::U8);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: tag_local,
-            rvalue: MirRValue::EnumTag { value: result_op.clone() },
-        });
-
-        let ok_block = self.builder.create_block();
-        let err_block = self.builder.create_block();
-        let merge_block = self.builder.create_block();
-
-        // Preserve Result type for correct codegen field offsets
-        let out_ty = result_ty.clone();
-        let out = self.builder.alloc_temp(out_ty.clone());
-
-        // tag=0 → Ok, tag=1 → Err
-        self.builder.terminate(MirTerminator::Branch {
-            cond: MirOperand::Local(tag_local),
-            then_block: err_block,
-            else_block: ok_block,
-        });
-
-        // Ok path: pass through unchanged
-        self.builder.switch_to_block(ok_block);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: out,
-            rvalue: MirRValue::Use(result_op.clone()),
-        });
-        self.builder.terminate(MirTerminator::Goto { target: merge_block });
-
-        // Err path: extract payload, wrap with constructor, re-wrap as Err
-        self.builder.switch_to_block(err_block);
-        let err_payload = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: err_payload,
-            rvalue: MirRValue::Field { base: result_op, field_index: 0, byte_offset: None, field_size: None },
-        });
-        // Wrap payload in variant: tag + payload
-        let constructor_tag = self.variant_tag(constructor_name);
-        let wrapped = self.builder.alloc_temp(MirType::Ptr);
-        self.builder.push_stmt(MirStmt::Store {
-            addr: wrapped,
-            offset: 0,
-            value: MirOperand::Constant(MirConst::Int(constructor_tag)),
-            store_size: None,
-        });
-        self.builder.push_stmt(MirStmt::Store {
-            addr: wrapped,
-            offset: 8,
-            value: MirOperand::Local(err_payload),
-            store_size: None,
-        });
-        // Re-wrap as Err(wrapped)
-        self.builder.push_stmt(MirStmt::Store {
-            addr: out,
-            offset: 0,
-            value: MirOperand::Constant(MirConst::Int(1)), // Err tag
-            store_size: None,
-        });
-        self.builder.push_stmt(MirStmt::Store {
-            addr: out,
-            offset: 8,
-            value: MirOperand::Local(wrapped),
-            store_size: None,
-        });
-        self.builder.terminate(MirTerminator::Goto { target: merge_block });
-
-        self.builder.switch_to_block(merge_block);
-        Ok((MirOperand::Local(out), out_ty))
-    }
-
-    /// Expand `Map.from([(k, v), ...])` into `Map.new()` + `Map.insert()` per pair.
-    /// Vec.from([a, b, c]) → store elements into stack array, call rask_vec_from_static
-    fn lower_vec_from_array(
-        &mut self,
-        elems: &[Expr],
-    ) -> Result<TypedOperand, LoweringError> {
-        // Lower elements and store into a stack array
-        let mut elem_ty = MirType::I64;
-        let mut lowered = Vec::new();
-        for (i, elem) in elems.iter().enumerate() {
-            let (op, ty) = self.lower_expr(elem)?;
-            if i == 0 {
-                elem_ty = ty;
-            }
-            lowered.push(op);
-        }
-        let elem_size = elem_ty.size();
-        let array_ty = MirType::Array {
-            elem: Box::new(elem_ty.clone()),
-            len: elems.len() as u32,
-        };
-        let arr_local = self.builder.alloc_temp(array_ty);
-        for (i, op) in lowered.into_iter().enumerate() {
-            self.builder.push_stmt(MirStmt::Store {
-                addr: arr_local,
-                offset: i as u32 * elem_size,
-                value: op,
-                store_size: None,
-            });
-        }
-
-        // rask_vec_from_static(data_ptr, count) → RaskVec*
-        let vec_local = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(vec_local),
-            func: FunctionRef::internal("rask_vec_from_static".to_string()),
-            args: vec![
-                MirOperand::Local(arr_local),
-                MirOperand::Constant(MirConst::Int(elems.len() as i64)),
-            ],
-        });
-        Ok((MirOperand::Local(vec_local), MirType::I64))
-    }
-
-    fn lower_map_from_pairs(
-        &mut self,
-        elems: &[Expr],
-    ) -> Result<TypedOperand, LoweringError> {
-        // Detect string keys from first pair
-        let has_string_keys = elems.first()
-            .and_then(|e| match &e.kind {
-                ExprKind::Tuple(parts) if parts.len() == 2 => {
-                    self.ctx.lookup_raw_type(parts[0].id)
-                        .map(|ty| matches!(ty, rask_types::Type::String))
-                },
-                _ => None,
-            })
-            .unwrap_or(false);
-
-        let ctor = if has_string_keys { "Map_new_string_keys" } else { "Map_new" };
-        let map_local = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(map_local),
-            func: FunctionRef::internal(ctor.to_string()),
-            args: vec![],
-        });
-
-        // Insert each (key, value) pair
-        for elem in elems {
-            let (key_op, val_op) = match &elem.kind {
-                ExprKind::Tuple(parts) if parts.len() == 2 => {
-                    let (k, _) = self.lower_expr(&parts[0])?;
-                    let (v, _) = self.lower_expr(&parts[1])?;
-                    (k, v)
-                }
-                _ => {
-                    // Non-tuple element — lower as value and skip
-                    let _ = self.lower_expr(elem)?;
-                    continue;
-                }
-            };
-            self.builder.push_stmt(MirStmt::Call {
-                dst: None,
-                func: FunctionRef::internal("Map_insert".to_string()),
-                args: vec![MirOperand::Local(map_local), key_op, val_op],
-            });
-        }
-
-        Ok((MirOperand::Local(map_local), MirType::I64))
-    }
-
-    /// Expand `json.encode(struct_val)` into a sequence of json_buf_* calls.
-    fn lower_json_encode_struct(
-        &mut self,
-        struct_op: MirOperand,
-        layout: StructLayout,
-    ) -> Result<TypedOperand, LoweringError> {
-        use rask_types::Type;
-
-        let buf = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(buf),
-            func: FunctionRef::internal("json_buf_new".to_string()),
-            args: vec![],
-        });
-
-        for (idx, field) in layout.fields.iter().enumerate() {
-            let field_val = self.builder.alloc_temp(MirType::I64);
-            self.builder.push_stmt(MirStmt::Assign {
-                dst: field_val,
-                rvalue: MirRValue::Field {
-                    base: struct_op.clone(),
-                    field_index: idx as u32,
-                    byte_offset: None,
-                    field_size: None,
-                },
-            });
-
-            // Nested struct: recursively encode and add as raw JSON
-            let nested_struct = match &field.ty {
-                Type::UnresolvedNamed(name) => self.ctx.find_struct(name).map(|(_, l)| l.clone()),
-                Type::UnresolvedGeneric { name, .. } => self.ctx.find_struct(name).map(|(_, l)| l.clone()),
-                _ => None,
-            };
-
-            if let Some(nested_layout) = nested_struct {
-                let (nested_json, _) = self.lower_json_encode_struct(
-                    MirOperand::Local(field_val),
-                    nested_layout,
-                )?;
-                self.builder.push_stmt(MirStmt::Call {
-                    dst: None,
-                    func: FunctionRef::internal("json_buf_add_raw".to_string()),
-                    args: vec![
-                        MirOperand::Local(buf),
-                        MirOperand::Constant(MirConst::String(field.name.clone())),
-                        nested_json,
-                    ],
-                });
-                continue;
-            }
-
-            let helper = match &field.ty {
-                Type::String => "json_buf_add_string",
-                Type::Bool => "json_buf_add_bool",
-                Type::F32 | Type::F64 => "json_buf_add_f64",
-                _ => "json_buf_add_i64",
-            };
-
-            self.builder.push_stmt(MirStmt::Call {
-                dst: None,
-                func: FunctionRef::internal(helper.to_string()),
-                args: vec![
-                    MirOperand::Local(buf),
-                    MirOperand::Constant(MirConst::String(field.name.clone())),
-                    MirOperand::Local(field_val),
-                ],
-            });
-        }
-
-        let result = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(result),
-            func: FunctionRef::internal("json_buf_finish".to_string()),
-            args: vec![MirOperand::Local(buf)],
-        });
-
-        Ok((MirOperand::Local(result), MirType::I64))
-    }
-
-    /// Expand `json.encode(vec)` into a loop that encodes each element and builds a JSON array.
-    fn lower_json_encode_vec(
-        &mut self,
-        vec_op: MirOperand,
-        elem_ty: Option<rask_types::Type>,
-    ) -> Result<TypedOperand, LoweringError> {
-        use rask_types::Type;
-
-        // arr_buf = json_buf_new_array()
-        let arr_buf = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(arr_buf),
-            func: FunctionRef::internal("json_buf_new_array".to_string()),
-            args: vec![],
-        });
-
-        // collection = vec_op (store in temp for repeated use)
-        let collection = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: collection,
-            rvalue: MirRValue::Use(vec_op),
-        });
-
-        // len = Vec_len(collection)
-        let len_local = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(len_local),
-            func: FunctionRef::internal("Vec_len".to_string()),
-            args: vec![MirOperand::Local(collection)],
-        });
-
-        // idx = 0
-        let idx = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: idx,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Int(0))),
-        });
-
-        let check_block = self.builder.create_block();
-        let body_block = self.builder.create_block();
-        let exit_block = self.builder.create_block();
-
-        self.builder.terminate(MirTerminator::Goto { target: check_block });
-
-        // check: idx < len
-        self.builder.switch_to_block(check_block);
-        let cond = self.builder.alloc_temp(MirType::Bool);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: cond,
-            rvalue: MirRValue::BinaryOp {
-                op: crate::operand::BinOp::Lt,
-                left: MirOperand::Local(idx),
-                right: MirOperand::Local(len_local),
-            },
-        });
-        self.builder.terminate(MirTerminator::Branch {
-            cond: MirOperand::Local(cond),
-            then_block: body_block,
-            else_block: exit_block,
-        });
-
-        // body: elem = Vec_get(collection, idx), encode, add to array
-        self.builder.switch_to_block(body_block);
-
-        // elem = Vec_get(collection, idx)
-        // Vec_get returns void*; codegen applies DerefResult to load the i64 value
-        let elem = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(elem),
-            func: FunctionRef::internal("Vec_get".to_string()),
-            args: vec![MirOperand::Local(collection), MirOperand::Local(idx)],
-        });
-
-        // Encode element based on type, then add to array buffer
-        let elem_ref = &elem_ty;
-        let nested_struct = match elem_ref {
-            Some(Type::UnresolvedNamed(name)) => self.ctx.find_struct(name).map(|(_, l)| l.clone()),
-            Some(Type::UnresolvedGeneric { name, .. }) => self.ctx.find_struct(name).map(|(_, l)| l.clone()),
-            Some(Type::Named(type_id)) => {
-                // Resolved type: look up name via type_names, then find struct
-                self.ctx.type_names.get(type_id)
-                    .and_then(|name| self.ctx.find_struct(name).map(|(_, l)| l.clone()))
-            }
-            _ => None,
-        };
-
-        if let Some(layout) = nested_struct {
-            // Struct element: encode to JSON string, add as raw
-            let (json_str, _) = self.lower_json_encode_struct(
-                MirOperand::Local(elem),
-                layout,
-            )?;
-            self.builder.push_stmt(MirStmt::Call {
-                dst: None,
-                func: FunctionRef::internal("json_buf_array_add_raw".to_string()),
-                args: vec![MirOperand::Local(arr_buf), json_str],
-            });
-        } else {
-            // Primitive element: pick the right array_add function
-            let helper = match elem_ref {
-                Some(Type::String) => "json_buf_array_add_string",
-                Some(Type::Bool) => "json_buf_array_add_bool",
-                Some(Type::F32) | Some(Type::F64) => "json_buf_array_add_f64",
-                _ => "json_buf_array_add_i64",
-            };
-            self.builder.push_stmt(MirStmt::Call {
-                dst: None,
-                func: FunctionRef::internal(helper.to_string()),
-                args: vec![MirOperand::Local(arr_buf), MirOperand::Local(elem)],
-            });
-        }
-
-        // idx = idx + 1
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: idx,
-            rvalue: MirRValue::BinaryOp {
-                op: crate::operand::BinOp::Add,
-                left: MirOperand::Local(idx),
-                right: MirOperand::Constant(MirConst::Int(1)),
-            },
-        });
-        self.builder.terminate(MirTerminator::Goto { target: check_block });
-
-        // exit: result = json_buf_finish_array(arr_buf)
-        self.builder.switch_to_block(exit_block);
-        let result = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(result),
-            func: FunctionRef::internal("json_buf_finish_array".to_string()),
-            args: vec![MirOperand::Local(arr_buf)],
-        });
-
-        Ok((MirOperand::Local(result), MirType::I64))
-    }
-
-    /// Expand `json.decode<T>(str)` into json_parse + field extraction + struct construction.
-    fn lower_json_decode_struct(
-        &mut self,
-        str_op: MirOperand,
-        layout: StructLayout,
-    ) -> Result<TypedOperand, LoweringError> {
-        use rask_types::Type;
-
-        // Parse JSON string into opaque object
-        let parsed = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(parsed),
-            func: FunctionRef::internal("json_parse".to_string()),
-            args: vec![str_op],
-        });
-
-        // Find struct type ID for result
-        let struct_id = self.ctx.find_struct(&layout.name)
-            .map(|(id, _)| StructLayoutId(id));
-        let struct_ty = struct_id
-            .map(MirType::Struct)
-            .unwrap_or(MirType::I64);
-
-        // Allocate struct and extract each field
-        let result = self.builder.alloc_temp(struct_ty.clone());
-        for (_idx, field) in layout.fields.iter().enumerate() {
-            let helper = match &field.ty {
-                Type::String => "json_get_string",
-                Type::Bool => "json_get_bool",
-                Type::F32 | Type::F64 => "json_get_f64",
-                _ => "json_get_i64",
-            };
-
-            let field_val = self.builder.alloc_temp(MirType::I64);
-            self.builder.push_stmt(MirStmt::Call {
-                dst: Some(field_val),
-                func: FunctionRef::internal(helper.to_string()),
-                args: vec![
-                    MirOperand::Local(parsed),
-                    MirOperand::Constant(MirConst::String(field.name.clone())),
-                ],
-            });
-
-            self.builder.push_stmt(MirStmt::Store {
-                addr: result,
-                offset: field.offset,
-                value: MirOperand::Local(field_val),
-                store_size: None,
-            });
-        }
-
-        Ok((MirOperand::Local(result), struct_ty))
-    }
-
-    /// Size in bytes for a MIR type (used for runtime allocation).
-    fn elem_size_for_type(&self, ty: &MirType) -> i64 {
-        match ty {
-            MirType::Bool | MirType::I8 | MirType::U8 => 1,
-            MirType::I16 | MirType::U16 => 2,
-            MirType::I32 | MirType::U32 | MirType::F32 | MirType::Char => 4,
-            MirType::I64 | MirType::U64 | MirType::F64 | MirType::Ptr
-            | MirType::String | MirType::FuncPtr(_) | MirType::Handle => 8,
-            MirType::Struct(StructLayoutId(id)) => {
-                self.ctx.struct_layouts.get(*id as usize)
-                    .map(|l| l.size as i64)
-                    .unwrap_or(8)
-            }
-            MirType::Enum(EnumLayoutId(id)) => {
-                self.ctx.enum_layouts.get(*id as usize)
-                    .map(|l| l.size as i64)
-                    .unwrap_or(8)
-            }
-            MirType::Array { elem, len } => self.elem_size_for_type(elem) * (*len as i64),
-            MirType::Tuple(_) | MirType::Slice(_) | MirType::Option(_)
-            | MirType::Result { .. } | MirType::Union(_)
-            | MirType::SimdVector { .. } | MirType::TraitObject { .. } => ty.size() as i64,
-            MirType::Void => 0,
-        }
-    }
-
-    // ═══════════════════════════════════════════════════════════
-    // Iterator chain recognition + inline expansion
-    // ═══════════════════════════════════════════════════════════
-
-    /// Walk a method call chain backward to find .iter() and collect adapters.
-    ///
-    /// vec.iter().filter(|x| p(x)).map(|x| f(x))
-    ///                                 ↑ start here, walk left
-    ///
-    /// Returns None if the chain doesn't end in .iter() or uses unsupported adapters.
-    pub(super) fn try_parse_iter_chain<'e>(&self, expr: &'e Expr) -> Option<super::IterChain<'e>> {
-        let mut adapters = Vec::new();
-        let mut current = expr;
-
-        loop {
-            match &current.kind {
-                ExprKind::MethodCall { object, method, args, .. } => {
-                    match method.as_str() {
-                        "iter" if args.is_empty() => {
-                            // Found the source — reverse adapters (we collected outside-in)
-                            adapters.reverse();
-                            return Some(super::IterChain {
-                                source: object,
-                                adapters,
-                            });
-                        }
-                        "filter" if args.len() == 1 => {
-                            if matches!(&args[0].expr.kind, ExprKind::Closure { .. }) {
-                                adapters.push(super::IterAdapter::Filter { closure: &args[0].expr });
-                                current = object;
-                            } else {
-                                return None;
-                            }
-                        }
-                        "map" if args.len() == 1 => {
-                            if matches!(&args[0].expr.kind, ExprKind::Closure { .. }) {
-                                adapters.push(super::IterAdapter::Map { closure: &args[0].expr });
-                                current = object;
-                            } else {
-                                return None;
-                            }
-                        }
-                        "take" if args.len() == 1 => {
-                            adapters.push(super::IterAdapter::Take { count: &args[0].expr });
-                            current = object;
-                        }
-                        "skip" if args.len() == 1 => {
-                            adapters.push(super::IterAdapter::Skip { count: &args[0].expr });
-                            current = object;
-                        }
-                        "enumerate" if args.is_empty() => {
-                            adapters.push(super::IterAdapter::Enumerate);
-                            current = object;
-                        }
-                        _ => return None, // Unknown adapter
-                    }
-                }
-                _ => return None, // Chain must be method calls all the way down
-            }
-        }
-    }
-
-    /// Clone function name for a type, or None if the type is Copy.
-    fn clone_fn_for_type(ty: &rask_types::Type) -> Option<&'static str> {
-        match ty {
-            rask_types::Type::String => Some("string_clone"),
-            rask_types::Type::UnresolvedNamed(n) if n == "string" => Some("string_clone"),
-            rask_types::Type::UnresolvedGeneric { name, .. } if name == "Vec" => Some("Vec_clone"),
-            rask_types::Type::UnresolvedGeneric { name, .. } if name == "Map" => Some("Map_clone"),
-            _ => None,
-        }
-    }
-
-    /// Emit inline clone for an enum value: shallow copy the full block,
-    /// then switch on the tag to deep-clone heap fields per variant.
-    fn lower_enum_clone(
-        &mut self,
-        layout: &rask_mono::EnumLayout,
-        src: &MirOperand,
-        obj_ty: MirType,
-    ) -> Result<TypedOperand, LoweringError> {
-        let result = self.builder.alloc_temp(obj_ty.clone());
-
-        // Shallow copy: copy each 8-byte word from src to result
-        let num_words = (layout.size as u32 + 7) / 8;
-        for i in 0..num_words {
-            let offset = i * 8;
-            let word = self.builder.alloc_temp(MirType::I64);
-            self.builder.push_stmt(MirStmt::Assign {
-                dst: word,
-                rvalue: MirRValue::Field {
-                    base: src.clone(),
-                    field_index: offset,
-                    byte_offset: None,
-                    field_size: None,
-                },
-            });
-            self.builder.push_stmt(MirStmt::Store {
-                addr: result,
-                offset,
-                value: MirOperand::Local(word),
-                store_size: None,
-            });
-        }
-
-        // Check if any variant has heap fields needing deep clone
-        let needs_switch = layout.variants.iter().any(|v| {
-            v.fields.iter().any(|f| Self::clone_fn_for_type(&f.ty).is_some())
-        });
-
-        if needs_switch {
-            // Read the tag
-            let tag = self.builder.alloc_temp(MirType::I64);
-            self.builder.push_stmt(MirStmt::Assign {
-                dst: tag,
-                rvalue: MirRValue::Field {
-                    base: MirOperand::Local(result),
-                    field_index: layout.tag_offset,
-                    byte_offset: None,
-                    field_size: None,
-                },
-            });
-
-            let exit_block = self.builder.create_block();
-            let mut cases = Vec::new();
-
-            for variant in &layout.variants {
-                let has_heap = variant.fields.iter().any(|f| Self::clone_fn_for_type(&f.ty).is_some());
-                if !has_heap {
-                    // No heap fields — shallow copy is sufficient, skip
-                    continue;
-                }
-                let vblock = self.builder.create_block();
-                cases.push((variant.tag as u64, vblock));
-
-                self.builder.switch_to_block(vblock);
-                for field in &variant.fields {
-                    if let Some(cfn) = Self::clone_fn_for_type(&field.ty) {
-                        let abs_offset = variant.payload_offset + field.offset;
-                        let field_val = self.builder.alloc_temp(MirType::I64);
-                        self.builder.push_stmt(MirStmt::Assign {
-                            dst: field_val,
-                            rvalue: MirRValue::Field {
-                                base: MirOperand::Local(result),
-                                field_index: abs_offset,
-                                byte_offset: None,
-                                field_size: None,
-                            },
-                        });
-                        let cloned = self.builder.alloc_temp(MirType::I64);
-                        self.builder.push_stmt(MirStmt::Call {
-                            dst: Some(cloned),
-                            func: FunctionRef::internal(cfn.to_string()),
-                            args: vec![MirOperand::Local(field_val)],
-                        });
-                        self.builder.push_stmt(MirStmt::Store {
-                            addr: result,
-                            offset: abs_offset,
-                            value: MirOperand::Local(cloned),
-                            store_size: None,
-                        });
-                    }
-                }
-                self.builder.terminate(MirTerminator::Goto { target: exit_block });
-            }
-
-            // Default case (variants with no heap fields) goes straight to exit
-            self.builder.terminate(MirTerminator::Switch {
-                value: MirOperand::Local(tag),
-                cases,
-                default: exit_block,
-            });
-
-            self.builder.switch_to_block(exit_block);
-        }
-
-        Ok((MirOperand::Local(result), obj_ty))
-    }
-
-    /// Try to handle an iterator terminal method (.collect, .fold, .any, etc.)
-    /// by recognizing the chain and emitting a fused loop.
-    ///
-    /// Returns Some if handled, None to fall through to regular method call.
-    fn try_lower_iter_terminal(
-        &mut self,
-        _full_expr: &Expr,
-        object: &Expr,
-        method: &str,
-        args: &[rask_ast::expr::CallArg],
-    ) -> Result<Option<TypedOperand>, LoweringError> {
-        match method {
-            "collect" if args.is_empty() => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    let result = self.lower_iter_collect(&chain)?;
-                    return Ok(Some(result));
-                }
-            }
-            "fold" if args.len() == 2 => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    let result = self.lower_iter_fold(&chain, &args[0].expr, &args[1].expr)?;
-                    return Ok(Some(result));
-                }
-            }
-            "any" if args.len() == 1 => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    if matches!(&args[0].expr.kind, ExprKind::Closure { .. }) {
-                        let result = self.lower_iter_any(&chain, &args[0].expr)?;
-                        return Ok(Some(result));
-                    }
-                }
-            }
-            "all" if args.len() == 1 => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    if matches!(&args[0].expr.kind, ExprKind::Closure { .. }) {
-                        let result = self.lower_iter_all(&chain, &args[0].expr)?;
-                        return Ok(Some(result));
-                    }
-                }
-            }
-            "count" if args.is_empty() => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    let result = self.lower_iter_count(&chain)?;
-                    return Ok(Some(result));
-                }
-            }
-            "sum" if args.is_empty() => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    let result = self.lower_iter_sum(&chain)?;
-                    return Ok(Some(result));
-                }
-            }
-            "find" if args.len() == 1 => {
-                if let Some(chain) = self.try_parse_iter_chain(object) {
-                    if matches!(&args[0].expr.kind, ExprKind::Closure { .. }) {
-                        let result = self.lower_iter_find(&chain, &args[0].expr)?;
-                        return Ok(Some(result));
-                    }
-                }
-            }
-            _ => {}
-        }
-        Ok(None)
-    }
-
-    /// Inline a closure body: substitute the closure parameter with a value
-    /// and lower the body expression. Used to fuse iterator adapters.
-    ///
-    /// |x| x * 2  +  arg_op  →  lower(x * 2) with x bound to arg_op
-    pub(super) fn inline_closure_body(
-        &mut self,
-        closure: &Expr,
-        arg_op: MirOperand,
-        arg_ty: MirType,
-    ) -> Result<TypedOperand, LoweringError> {
-        if let ExprKind::Closure { params, body, .. } = &closure.kind {
-            if let Some(param) = params.first() {
-                let param_name = &param.name;
-                // Save existing binding
-                let saved = self.locals.remove(param_name);
-                // Bind parameter to the argument value
-                let param_local = self.builder.alloc_local(param_name.clone(), arg_ty.clone());
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: param_local,
-                    rvalue: MirRValue::Use(arg_op),
-                });
-                self.locals.insert(param_name.clone(), (param_local, arg_ty));
-                // Lower closure body
-                let result = self.lower_expr(body)?;
-                // Restore previous binding
-                self.locals.remove(param_name);
-                if let Some(prev) = saved {
-                    self.locals.insert(param_name.clone(), prev);
-                }
-                return Ok(result);
-            }
-        }
-        Err(LoweringError::InvalidConstruct("expected closure".to_string()))
-    }
-
-    /// Set up the index loop infrastructure for an iterator chain.
-    /// Returns (collection_local, idx_local, len_local, elem_ty, is_array, check_block, body_block, inc_block, exit_block).
-    pub(super) fn setup_iter_chain_loop(
-        &mut self,
-        chain: &super::IterChain<'_>,
-    ) -> Result<IterLoopSetup, LoweringError> {
-        let (source_op, source_ty) = self.lower_expr(chain.source)?;
-        let is_array = matches!(&source_ty, MirType::Array { .. });
-        let (array_len, array_elem_size) = match &source_ty {
-            MirType::Array { elem, len } => (Some(*len), Some(elem.size())),
-            _ => (None, None),
-        };
-
-        let collection = self.builder.alloc_temp(source_ty.clone());
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: collection,
-            rvalue: MirRValue::Use(source_op),
-        });
-
-        // len
-        let len_local = self.builder.alloc_temp(MirType::I64);
-        if let Some(arr_len) = array_len {
-            self.builder.push_stmt(MirStmt::Assign {
-                dst: len_local,
-                rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Int(arr_len as i64))),
-            });
-        } else {
-            self.builder.push_stmt(MirStmt::Call {
-                dst: Some(len_local),
-                func: FunctionRef::internal("Vec_len".to_string()),
-                args: vec![MirOperand::Local(collection)],
-            });
-        }
-
-        // Process Skip/Take adapters to adjust start/end bounds
-        let mut start_val: Option<MirOperand> = None;
-        let mut end_op = MirOperand::Local(len_local);
-
-        for adapter in &chain.adapters {
-            match adapter {
-                super::IterAdapter::Skip { count } => {
-                    let (skip_op, _) = self.lower_expr(count)?;
-                    start_val = Some(skip_op);
-                }
-                super::IterAdapter::Take { count } => {
-                    let (take_op, _) = self.lower_expr(count)?;
-                    // end = min(start + take, len) — simplified to start + take
-                    if let Some(ref start) = start_val {
-                        let adjusted = self.builder.alloc_temp(MirType::I64);
-                        self.builder.push_stmt(MirStmt::Assign {
-                            dst: adjusted,
-                            rvalue: MirRValue::BinaryOp {
-                                op: crate::operand::BinOp::Add,
-                                left: start.clone(),
-                                right: take_op,
-                            },
-                        });
-                        end_op = MirOperand::Local(adjusted);
-                    } else {
-                        end_op = take_op;
-                    }
-                }
-                _ => {} // Filter/Map/Enumerate handled inside the loop body
-            }
-        }
-
-        let idx = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: idx,
-            rvalue: MirRValue::Use(start_val.unwrap_or(MirOperand::Constant(MirConst::Int(0)))),
-        });
-
-        let check_block = self.builder.create_block();
-        let body_block = self.builder.create_block();
-        let inc_block = self.builder.create_block();
-        let exit_block = self.builder.create_block();
-
-        self.builder.terminate(MirTerminator::Goto { target: check_block });
-
-        // check: idx < end
-        self.builder.switch_to_block(check_block);
-        let cond = self.builder.alloc_temp(MirType::Bool);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: cond,
-            rvalue: MirRValue::BinaryOp {
-                op: crate::operand::BinOp::Lt,
-                left: MirOperand::Local(idx),
-                right: end_op,
-            },
-        });
-        self.builder.terminate(MirTerminator::Branch {
-            cond: MirOperand::Local(cond),
-            then_block: body_block,
-            else_block: exit_block,
-        });
-
-        // body: load element
-        self.builder.switch_to_block(body_block);
-        let elem_ty = self.extract_iterator_elem_type(chain.source)
-            .unwrap_or(MirType::I64);
-        let elem_local = self.builder.alloc_temp(elem_ty.clone());
-        if is_array {
-            self.builder.push_stmt(MirStmt::Assign {
-                dst: elem_local,
-                rvalue: MirRValue::ArrayIndex {
-                    base: MirOperand::Local(collection),
-                    index: MirOperand::Local(idx),
-                    elem_size: array_elem_size.unwrap_or(8),
-                },
-            });
-        } else {
-            self.builder.push_stmt(MirStmt::Call {
-                dst: Some(elem_local),
-                func: FunctionRef::internal("Vec_get".to_string()),
-                args: vec![MirOperand::Local(collection), MirOperand::Local(idx)],
-            });
-        }
-
-        Ok(IterLoopSetup {
-            idx,
-            elem_local,
-            elem_ty,
-            inc_block,
-            exit_block,
-            check_block,
-        })
-    }
-
-    /// Apply filter/map/enumerate adapters inside a loop body.
-    /// Returns the final (operand, type) after all adapters.
-    /// For filter adapters, emits a branch that skips to inc_block on false.
-    pub(super) fn apply_iter_adapters(
-        &mut self,
-        chain: &super::IterChain<'_>,
-        elem_op: MirOperand,
-        elem_ty: MirType,
-        inc_block: crate::BlockId,
-        idx: crate::LocalId,
-    ) -> Result<TypedOperand, LoweringError> {
-        let mut current_op = elem_op;
-        let mut current_ty = elem_ty;
-        // enumerate_idx is used below when building the tuple
-
-        for adapter in &chain.adapters {
-            match adapter {
-                super::IterAdapter::Filter { closure } => {
-                    let (pred_op, _) = self.inline_closure_body(closure, current_op.clone(), current_ty.clone())?;
-                    // Continue block for elements that pass the filter
-                    let pass_block = self.builder.create_block();
-                    self.builder.terminate(MirTerminator::Branch {
-                        cond: pred_op,
-                        then_block: pass_block,
-                        else_block: inc_block,
-                    });
-                    self.builder.switch_to_block(pass_block);
-                }
-                super::IterAdapter::Map { closure } => {
-                    let (mapped_op, mapped_ty) = self.inline_closure_body(closure, current_op, current_ty)?;
-                    current_op = mapped_op;
-                    current_ty = mapped_ty;
-                }
-                super::IterAdapter::Enumerate => {
-                    // Build (index, element) tuple on the stack
-                    let tuple_ty = MirType::Tuple(vec![MirType::I64, current_ty.clone()]);
-                    let tuple_local = self.builder.alloc_temp(tuple_ty.clone());
-                    // field 0: index
-                    self.builder.push_stmt(MirStmt::Store {
-                        addr: tuple_local,
-                        offset: 0,
-                        value: MirOperand::Local(idx),
-                        store_size: None,
-                    });
-                    // field 1: element (offset 8, aligned)
-                    self.builder.push_stmt(MirStmt::Store {
-                        addr: tuple_local,
-                        offset: 8,
-                        value: current_op,
-                        store_size: None,
-                    });
-                    current_op = MirOperand::Local(tuple_local);
-                    current_ty = tuple_ty;
-                }
-                super::IterAdapter::Skip { .. } | super::IterAdapter::Take { .. } => {
-                    // Already handled in setup (start/end bounds)
-                }
-            }
-        }
-
-        Ok((current_op, current_ty))
-    }
-
-    /// Emit the increment block: idx += 1, goto check
-    pub(super) fn emit_iter_increment(
-        &mut self,
-        idx: crate::LocalId,
-        inc_block: crate::BlockId,
-        check_block: crate::BlockId,
-    ) {
-        self.builder.switch_to_block(inc_block);
-        let incremented = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: incremented,
-            rvalue: MirRValue::BinaryOp {
-                op: crate::operand::BinOp::Add,
-                left: MirOperand::Local(idx),
-                right: MirOperand::Constant(MirConst::Int(1)),
-            },
-        });
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: idx,
-            rvalue: MirRValue::Use(MirOperand::Local(incremented)),
-        });
-        self.builder.terminate(MirTerminator::Goto { target: check_block });
-    }
-
-    /// .collect() — fused loop that pushes each result into a new Vec.
-    fn lower_iter_collect(
-        &mut self,
-        chain: &super::IterChain<'_>,
-    ) -> Result<TypedOperand, LoweringError> {
-        // Create result Vec
-        let result_vec = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Call {
-            dst: Some(result_vec),
-            func: FunctionRef::internal("Vec_new".to_string()),
-            args: vec![],
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let (final_op, _) = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        // Push into result Vec
-        self.builder.push_stmt(MirStmt::Call {
-            dst: None,
-            func: FunctionRef::internal("Vec_push".to_string()),
-            args: vec![MirOperand::Local(result_vec), final_op],
-        });
-        self.builder.terminate(MirTerminator::Goto { target: setup.inc_block });
-
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(result_vec), MirType::I64))
-    }
-
-    /// .fold(init, |acc, x| body) — fused loop with accumulator.
-    fn lower_iter_fold(
-        &mut self,
-        chain: &super::IterChain<'_>,
-        init: &Expr,
-        closure: &Expr,
-    ) -> Result<TypedOperand, LoweringError> {
-        let (init_op, init_ty) = self.lower_expr(init)?;
-        let acc = self.builder.alloc_temp(init_ty.clone());
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: acc,
-            rvalue: MirRValue::Use(init_op),
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let (final_op, final_ty) = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        // Inline the fold closure with two args: (acc, elem)
-        if let ExprKind::Closure { params, body, .. } = &closure.kind {
-            if params.len() == 2 {
-                let acc_name = &params[0].name;
-                let elem_name = &params[1].name;
-
-                let saved_acc = self.locals.remove(acc_name);
-                let saved_elem = self.locals.remove(elem_name);
-
-                let acc_param = self.builder.alloc_local(acc_name.clone(), init_ty.clone());
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: acc_param,
-                    rvalue: MirRValue::Use(MirOperand::Local(acc)),
-                });
-                self.locals.insert(acc_name.clone(), (acc_param, init_ty.clone()));
-
-                let elem_param = self.builder.alloc_local(elem_name.clone(), final_ty.clone());
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: elem_param,
-                    rvalue: MirRValue::Use(final_op),
-                });
-                self.locals.insert(elem_name.clone(), (elem_param, final_ty));
-
-                // A continue block for after the inlined body finishes normally
-                let after_body = self.builder.create_block();
-
-                // Redirect `return` inside the closure to assign to acc and continue the loop
-                let saved_return_target = self.inline_return_target.take();
-                self.inline_return_target = Some((acc, setup.inc_block));
-
-                let (result_op, _) = self.lower_expr(body)?;
-
-                self.inline_return_target = saved_return_target;
-
-                // Normal (non-return) path: assign result to acc
-                self.builder.push_stmt(MirStmt::Assign {
-                    dst: acc,
-                    rvalue: MirRValue::Use(result_op),
-                });
-                self.builder.terminate(MirTerminator::Goto { target: after_body });
-
-                self.builder.switch_to_block(after_body);
-
-                self.locals.remove(acc_name);
-                self.locals.remove(elem_name);
-                if let Some(prev) = saved_acc { self.locals.insert(acc_name.clone(), prev); }
-                if let Some(prev) = saved_elem { self.locals.insert(elem_name.clone(), prev); }
-            }
-        }
-
-        self.builder.terminate(MirTerminator::Goto { target: setup.inc_block });
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(acc), init_ty))
-    }
-
-    /// .any(|x| pred) — fused loop, short-circuit on first true.
-    fn lower_iter_any(
-        &mut self,
-        chain: &super::IterChain<'_>,
-        predicate: &Expr,
-    ) -> Result<TypedOperand, LoweringError> {
-        let result = self.builder.alloc_temp(MirType::Bool);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: result,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Bool(false))),
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let (final_op, final_ty) = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        let (pred_op, _) = self.inline_closure_body(predicate, final_op, final_ty)?;
-        let found_block = self.builder.create_block();
-        self.builder.terminate(MirTerminator::Branch {
-            cond: pred_op,
-            then_block: found_block,
-            else_block: setup.inc_block,
-        });
-
-        // Found — set result true and exit
-        self.builder.switch_to_block(found_block);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: result,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Bool(true))),
-        });
-        self.builder.terminate(MirTerminator::Goto { target: setup.exit_block });
-
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(result), MirType::Bool))
-    }
-
-    /// .all(|x| pred) — fused loop, short-circuit on first false.
-    fn lower_iter_all(
-        &mut self,
-        chain: &super::IterChain<'_>,
-        predicate: &Expr,
-    ) -> Result<TypedOperand, LoweringError> {
-        let result = self.builder.alloc_temp(MirType::Bool);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: result,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Bool(true))),
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let (final_op, final_ty) = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        let (pred_op, _) = self.inline_closure_body(predicate, final_op, final_ty)?;
-        let fail_block = self.builder.create_block();
-        self.builder.terminate(MirTerminator::Branch {
-            cond: pred_op,
-            then_block: setup.inc_block,
-            else_block: fail_block,
-        });
-
-        // Failed — set result false and exit
-        self.builder.switch_to_block(fail_block);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: result,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Bool(false))),
-        });
-        self.builder.terminate(MirTerminator::Goto { target: setup.exit_block });
-
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(result), MirType::Bool))
-    }
-
-    /// .count() — fused loop counting elements that pass filters.
-    fn lower_iter_count(
-        &mut self,
-        chain: &super::IterChain<'_>,
-    ) -> Result<TypedOperand, LoweringError> {
-        let counter = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: counter,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Int(0))),
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let _ = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        // Increment counter
-        let incremented = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: incremented,
-            rvalue: MirRValue::BinaryOp {
-                op: crate::operand::BinOp::Add,
-                left: MirOperand::Local(counter),
-                right: MirOperand::Constant(MirConst::Int(1)),
-            },
-        });
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: counter,
-            rvalue: MirRValue::Use(MirOperand::Local(incremented)),
-        });
-
-        self.builder.terminate(MirTerminator::Goto { target: setup.inc_block });
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(counter), MirType::I64))
-    }
-
-    /// .sum() — fused loop accumulating with Add.
-    fn lower_iter_sum(
-        &mut self,
-        chain: &super::IterChain<'_>,
-    ) -> Result<TypedOperand, LoweringError> {
-        let acc = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: acc,
-            rvalue: MirRValue::Use(MirOperand::Constant(MirConst::Int(0))),
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let (final_op, _) = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        let sum = self.builder.alloc_temp(MirType::I64);
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: sum,
-            rvalue: MirRValue::BinaryOp {
-                op: crate::operand::BinOp::Add,
-                left: MirOperand::Local(acc),
-                right: final_op,
-            },
-        });
-        self.builder.push_stmt(MirStmt::Assign {
-            dst: acc,
-            rvalue: MirRValue::Use(MirOperand::Local(sum)),
-        });
-
-        self.builder.terminate(MirTerminator::Goto { target: setup.inc_block });
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(acc), MirType::I64))
-    }
-
-    /// .find(|x| pred) — fused loop, return Some on first match, None otherwise.
-    fn lower_iter_find(
-        &mut self,
-        chain: &super::IterChain<'_>,
-        predicate: &Expr,
-    ) -> Result<TypedOperand, LoweringError> {
-        // Result: Option<I64> (tag at offset 0, payload at offset 8)
-        let result = self.builder.alloc_temp(MirType::Option(Box::new(MirType::I64)));
-        // Start as None (tag = 1)
-        self.builder.push_stmt(MirStmt::Store {
-            addr: result,
-            offset: 0,
-            value: MirOperand::Constant(MirConst::Int(1)),
-            store_size: None,
-        });
-
-        let setup = self.setup_iter_chain_loop(chain)?;
-        let (final_op, final_ty) = self.apply_iter_adapters(
-            chain, MirOperand::Local(setup.elem_local), setup.elem_ty,
-            setup.inc_block, setup.idx,
-        )?;
-
-        let (pred_op, _) = self.inline_closure_body(predicate, final_op.clone(), final_ty)?;
-        let found_block = self.builder.create_block();
-        self.builder.terminate(MirTerminator::Branch {
-            cond: pred_op,
-            then_block: found_block,
-            else_block: setup.inc_block,
-        });
-
-        // Found — set result to Some(elem)
-        self.builder.switch_to_block(found_block);
-        self.builder.push_stmt(MirStmt::Store {
-            addr: result,
-            offset: 0,
-            value: MirOperand::Constant(MirConst::Int(0)), // Some tag
-            store_size: None,
-        });
-        self.builder.push_stmt(MirStmt::Store {
-            addr: result,
-            offset: 8,
-            value: final_op,
-            store_size: None,
-        });
-        self.builder.terminate(MirTerminator::Goto { target: setup.exit_block });
-
-        self.emit_iter_increment(setup.idx, setup.inc_block, setup.check_block);
-        self.builder.switch_to_block(setup.exit_block);
-        Ok((MirOperand::Local(result), MirType::Option(Box::new(MirType::I64))))
-    }
-}
-
-/// Internal state for iterator chain loop setup.
-pub(super) struct IterLoopSetup {
-    pub(super) idx: crate::LocalId,
-    pub(super) elem_local: crate::LocalId,
-    pub(super) elem_ty: MirType,
-    pub(super) inc_block: crate::BlockId,
-    pub(super) exit_block: crate::BlockId,
-    pub(super) check_block: crate::BlockId,
-}
+} // end impl MirLowerer
